requirements.txt
fastapi
pydantic
pydantic_settings
uvicorn
pybars3
email-validator
